+++ knox/views.py
@@ -8,7 +8,7 @@ from rest_framework.settings import api_settings
 from rest_framework.views import APIView
 
 from knox.auth import TokenAuthentication
-from knox.models import get_token_model
+from knox.models import AuthToken
 from knox.settings import knox_settings
 
 
@@ -22,9 +22,6 @@ class LoginView(APIView):
     def get_token_ttl(self):
         return knox_settings.TOKEN_TTL
 
-    def get_token_prefix(self):
-        return knox_settings.TOKEN_PREFIX
-
     def get_token_limit_per_user(self):
         return knox_settings.TOKEN_LIMIT_PER_USER
 
@@ -37,12 +34,6 @@ class LoginView(APIView):
     def format_expiry_datetime(self, expiry):
         datetime_format = self.get_expiry_datetime_format()
         return DateTimeField(format=datetime_format).to_representation(expiry)
-
-    def create_token(self):
-        token_prefix = self.get_token_prefix()
-        return get_token_model().objects.create(
-            user=self.request.user, expiry=self.get_token_ttl(), prefix=token_prefix
-        )
 
     def get_post_response_data(self, request, token, instance):
         UserSerializer = self.get_user_serializer_class()
@@ -58,10 +49,6 @@ class LoginView(APIView):
             ).data
         return data
 
-    def get_post_response(self, request, token, instance):
-        data = self.get_post_response_data(request, token, instance)
-        return Response(data)
-
     def post(self, request, format=None):
         token_limit_per_user = self.get_token_limit_per_user()
         if token_limit_per_user is not None:
@@ -72,24 +59,23 @@ class LoginView(APIView):
                     {"error": "Maximum amount of tokens allowed per user exceeded."},
                     status=status.HTTP_403_FORBIDDEN
                 )
-        instance, token = self.create_token()
+        token_ttl = self.get_token_ttl()
+        instance, token = AuthToken.objects.create(request.user, token_ttl)
         user_logged_in.send(sender=request.user.__class__,
                             request=request, user=request.user)
-        return self.get_post_response(request, token, instance)
+        data = self.get_post_response_data(request, token, instance)
+        return Response(data)
 
 
 class LogoutView(APIView):
     authentication_classes = (TokenAuthentication,)
     permission_classes = (IsAuthenticated,)
 
-    def get_post_response(self, request):
-        return Response(None, status=status.HTTP_204_NO_CONTENT)
-
     def post(self, request, format=None):
         request._auth.delete()
         user_logged_out.send(sender=request.user.__class__,
                              request=request, user=request.user)
-        return self.get_post_response(request)
+        return Response(None, status=status.HTTP_204_NO_CONTENT)
 
 
 class LogoutAllView(APIView):
@@ -100,11 +86,8 @@ class LogoutAllView(APIView):
     authentication_classes = (TokenAuthentication,)
     permission_classes = (IsAuthenticated,)
 
-    def get_post_response(self, request):
-        return Response(None, status=status.HTTP_204_NO_CONTENT)
-
     def post(self, request, format=None):
         request.user.auth_token_set.all().delete()
         user_logged_out.send(sender=request.user.__class__,
                              request=request, user=request.user)
-        return self.get_post_response(request)
+        return Response(None, status=status.HTTP_204_NO_CONTENT)