+++ knox_project/settings.py
@@ -49,9 +49,7 @@ DATABASES = {
 LANGUAGE_CODE = 'en-us'
 TIME_ZONE = 'UTC'
 USE_I18N = True
-USE_L10N = True  # Deprecated since django 4.0.
+USE_L10N = True
 USE_TZ = True
 
 STATIC_URL = '/static/'
-
-KNOX_TOKEN_MODEL = 'knox.AuthToken'